chore(ci): drop Node.js 8
BREAKING CHANGE:  This library will no longer be tested on Node.js 8 and as such will not be guaranteed to work on this version of Node.js. We support LTS versions of Node.js as they are specified here: https://github.com/nodejs/Release#release-schedule

Note: Node.js 8 technically does not reach end of life until December 31st, 2019, but we are opting to drop support right now in order to align the release with #52.